Another fix to 32-bit iojs binary builds.
The iojs tarballs have begun to use `ia32` in their directory name, instead
of `x86`.

+++ scripts/publish_x86.sh
@@ -23,14 +23,18 @@ if [[ ${NODE_VERSION:0:4} == 'iojs' ]]; then
     wget https://iojs.org/download/release/${NVER}/iojs-${NVER}-${platform}-x86.tar.gz
     tar xf iojs-${NVER}-${platform}-x86.tar.gz
     # enable 32 bit iojs
-    export PATH=$(pwd)/iojs-${NVER}-${platform}-x86/bin:$PATH
+    export PATH=$(pwd)/iojs-${NVER}-${platform}-x86/bin:$(pwd)/iojs-${NVER}-${platform}-ia32/bin:$PATH
 else
     wget http://nodejs.org/dist/${NVER}/node-${NVER}-${platform}-x86.tar.gz
     tar xf node-${NVER}-${platform}-x86.tar.gz
     # enable 32 bit node
-    export PATH=$(pwd)/node-${NVER}-${platform}-x86/bin:$PATH
+    export PATH=$(pwd)/node-${NVER}-${platform}-x86/bin:$(pwd)/node-${NVER}-${platform}-ia32/bin:$PATH
 fi
 node -e "console.log(process.arch,process.execPath)"
+if [ `node -e "console.log(process.arch)"` != "ia32" ]; then
+    echo "Can't built 32-bit binaries."
+    exit 1
+fi
 # install 32 bit compiler toolchain and X11
 # test source compile in 32 bit mode
 CC=gcc-4.6 CXX=g++-4.6 npm install --build-from-source
